admin panel crud lengkap
tinggal integrasi ke front end

+++ laravel/app/Http/Controllers/StaffController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Staff;
+use Illuminate\Support\Facades\Storage;
 
 class StaffController extends Controller
 {
@@ -26,11 +27,15 @@ class StaffController extends Controller
             'foto' => 'image|mimes:jpeg,png,jpg|nullable',
         ]);
 
+        $fileName = null;
+
         if ($request->hasFile('foto')) {
             $file = $request->file('foto');
             $extension = $file->getClientOriginalExtension();
             $fileName = hash('md5', time() . $file->getClientOriginalName()) . '.' . $extension;
-            $file->move(public_path('imgstaff'), $fileName);
+
+            // Simpan di storage/app/public/imgstaff
+            $file->storeAs('public/imgstaff', $fileName);
         }
 
         Staff::create([
@@ -40,6 +45,12 @@ class StaffController extends Controller
         ]);
 
         return redirect('/staff')->with('success', 'Staff berhasil ditambahkan.');
+    }
+    public function show(string $id)
+    {
+        $staff = Staff::findOrFail($id);
+
+        return view('staff.show', compact('staff'));
     }
 
     public function edit($id)
@@ -53,22 +64,43 @@ class StaffController extends Controller
         $request->validate([
             'nama' => 'required',
             'jabatan' => 'required',
+            'foto' => 'image|mimes:jpeg,png,jpg|nullable',
         ]);
 
         $staff = Staff::find($id);
+
+        // Hapus gambar lama jika ada
+        if ($staff->foto) {
+            Storage::delete('public/imgstaff/' . $staff->foto);
+        }
+
+        $fileName = $staff->foto;
+
+        if ($request->hasFile('foto')) {
+            $file = $request->file('foto');
+            $extension = $file->getClientOriginalExtension();
+            $fileName = hash('md5', time() . $file->getClientOriginalName()) . '.' . $extension;
+
+            // Simpan di storage/app/public/imgstaff
+            $file->storeAs('public/imgstaff', $fileName);
+        }
+
         $staff->update([
             'nama' => $request->nama,
             'jabatan' => $request->jabatan,
+            'foto' => $fileName,
         ]);
 
-        return redirect()->route('staffs.index')->with('success', 'Data staff berhasil diperbarui.');
+        return redirect('/staff')->with('success', 'Data Staff berhasil diperbarui.');
     }
-
     public function destroy($id)
     {
         $staff = Staff::find($id);
+        if (!$staff) {
+            return redirect()->route('staff')->with('error', 'Data staff tidak ditemukan.');
+        }
         $staff->delete();
 
-        return redirect()->route('staffs.index')->with('success', 'Data staff berhasil dihapus.');
+        return redirect()->route('staff')->with('success', 'Data staff berhasil dihapus.');
     }
 }
